Some basic Database manipulation
The Makefile template has been updated to remove a new project's
dependency on lair build version --increment and to incorporate this new
metadata method into new project makefiles.

Added a middleware directory

Started scaffolding out default auth middleware in project gen

Instead of panics/Printlns, most of the output is now channeled through
Logrus.

Updated Makefile to use the air tool with make run (giving me hot
reload).

Added some macros into the seed functionality. The two macros that I'm
specifically interested in right now is $PASSWORD (for creating a
bycrpt-hashed password) and $UUID (for creating a random UUID).

$UUID hasn't been implemented here.

$PASSWORD is implemented but it's either not working or I'm not reading
bytes out of the database correctly. When bcrypt tries to compare
passwords against this value it's complaining that the secret is "too
short to be a bcrypted password." So, there's that.

Converted project from Docopt to Cobra.