fix: Updated name of migrations directory
In earlier concepts of Lair, the migrations directory lived under
db/migrations. For various reasions, I now store database-related code
in Gin projects under database/migrations. This patch updates that
decision in the code.

It might make sense to make this configurable in the future. For now,
I'm going to live with this hardwiring.